Реализован класс ReadOnlyStream
Дополнительно изменен IReadOnlyStream
Дополнительно изменена структура проекта
Дополнены саммари
Почищен gitingore
