+++ .gitignore
@@ -1,14 +1,3 @@
-##игнорируем сам .gitignore
-#.gitignore
-
-### игнорируем конкретный тип файлов
-## *.html
-###… кроме определенного файла
-## !special.html
-### можно использовать рег.выражения.
-## *.[ao]
-
-
 *.suo
 *.user
 .vs/
